# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Team Task Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /users/{id}/role:
    patch:
      operationId: updateUserRole
      tags: [users]
      summary: Update a user's app-level role (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRoleBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "403":
          description: Forbidden

  # Projects
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks (optionally filter by project)
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
        - name: assigneeId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [todo, in_progress, in_review, done]
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high, urgent]
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get task details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Dashboard / Analytics
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: string
        name:
          type: string
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, member]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clerkId
        - email
        - name
        - role
        - createdAt
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        avatarUrl:
          type: ["string", "null"]
    UpdateUserRoleBody:
      type: object
      properties:
        role:
          type: string
          enum: [admin, member]
      required:
        - role
...
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [todo, in_progress, in_review, done]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        projectId:
          type: integer
        assigneeId:
          type: ["integer", "null"]
        creatorId:
          type: integer
        dueDate:
          type: ["string", "null"]
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
        project:
          $ref: "#/components/schemas/ProjectSummary"
        assignee:
          oneOf:
            - $ref: "#/components/schemas/User"
            - type: "null"
        creator:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - status
        - priority
        - projectId
        - creatorId
        - project
        - creator
        - createdAt
        - updatedAt
    ProjectSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: string
      required:
        - id
        - name
        - color
    CreateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [todo, in_progress, in_review, done]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        projectId:
          type: integer
        assigneeId:
          type: ["integer", "null"]
        dueDate:
          type: ["string", "null"]
          format: date-time
      required:
        - title
        - projectId
        - status
        - priority
    UpdateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [todo, in_progress, in_review, done]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        assigneeId:
          type: ["integer", "null"]
        dueDate:
          type: ["string", "null"]
          format: date-time
...